finalizado controller User e realizado os devidos Testes

+++ api/src/controller/UserController.js
@@ -3,16 +3,21 @@ const User = require('../models/User');
 class UserController {
     async store(req, res) {
         try {
-            const { name, email, password } = req.body;
-            const user = await User.create({
-                name, email, password
-            });
+            const data = {
+                name: req.body.name,
+                email: req.body.email,
+                phone: req.body.phone,
+                password: req.body.password
+            }           
+
+            const user = await User.create(data);
             return res.status(201).json(user);
+
         } catch (error) {
+            console.error(error);
             return res.status(400).json({
                 'status': '400',
-                'Failure': error.name,
-                'error': error.original.sql,
+                'error': error.error
             });
         }
     }
@@ -23,14 +28,14 @@ class UserController {
         } catch (error) {
             return res.status(400).json({
                 'status': '400',
-                'Failure': error.name,
-                'error': error.original.sql,
+                'error': 'no items found',
             });
         }
     }
     async findById(req, res) {
         try {
             const { id } = req.params;
+            if (!id) throw new Validation('id is required');
             const users = await User.findAll({
                 where: {
                     id: id
@@ -40,8 +45,7 @@ class UserController {
         } catch (error) {
             return res.status(400).json({
                 'status': '400',
-                'Failure': error.name,
-                'error': error.original.sql,
+                'error': error.error
             });
         }
     }
@@ -49,6 +53,7 @@ class UserController {
     async delete(req, res) {
         try {
             const { id } = req.params;
+            if (!id) throw new Validation('id is required');
             const users = await User.destroy({
                 where: {
                     id: id
@@ -58,8 +63,7 @@ class UserController {
         } catch (error) {
             return res.status(400).json({
                 'status': '400',
-                'Failure': error.name,
-                'error': error.original.sql,
+                'error': error.error
             });
         }
     }
@@ -69,20 +73,23 @@ class UserController {
             const data = {
                 name: req.body.name,
                 email: req.body.email,
+                phone: req.body.phone,
                 password: req.body.password,
             }
+           
             const { id } = req.params;
+            if (!id) throw new Validation('id is required');
             const user = await User.update(data, {
                 where: {
                     id: id
-                }
+                },
+                individualHooks: true
             });
             return res.status(200).json(user);
         } catch (error) {
             return res.status(400).json({
                 'status': '400',
-                'Failure': error.name,
-                'error': error.original.sql,
+                'error': error.error
             });
         }
     }
